Corrected data type to float64

+++ models/staging/erp_northwind/stg_erp_northwind__products.sql
@@ -9,7 +9,7 @@ with
             cast(supplier_id as int) as supplier_id,
             cast(category_id as int) as category_id,
             cast(quantity_per_unit as string) as quantity_per_unit,
-            cast(unit_price as float) as unit_price,
+            cast(unit_price as float64) as unit_price,
             cast(units_in_stock as int) as units_in_stock,
             cast(units_on_order as int) as units_on_order,
             cast(reorder_level as int) as reorder_level,
